avx512 kernel

+++ benchmark/defines.h
@@ -3,8 +3,8 @@
 using FloatComplex = std::complex<float>;
 using DoubleComplex = std::complex<double>;
 
-typedef float floatType;
-//typedef double floatType;
+//typedef float floatType;
+typedef double floatType;
 //typedef DoubleComplex floatType;
 //typedef FloatComplex floatType;
